+++ src/main/java/roomescape/reservation/service/PaymentClient.java
@@ -6,62 +6,18 @@ import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.Base64;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.ssl.SslBundles;
 import org.springframework.boot.web.client.ClientHttpRequestFactories;
 import org.springframework.boot.web.client.ClientHttpRequestFactorySettings;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.MediaType;
 import org.springframework.http.client.ClientHttpRequestFactory;
-import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 import roomescape.global.exception.IllegalRequestException;
 import roomescape.global.exception.InternalServerException;
 import roomescape.global.exception.PaymentErrorResponse;
 import roomescape.reservation.dto.PaymentConfirmRequest;
 
-@Component
-public class PaymentClient {
-
-    private static final String BASE_URL = "https://api.tosspayments.com";
-    private static final String CONFIRM_URI = "/v1/payments/confirm";
-    private static final String AUTH_TYPE = "Basic ";
-
-    @Value("${third-party-api.payment.secret-key}")
-    private String secretKey;
-    private final RestClient restClient;
-    private final ObjectMapper objectMapper = new ObjectMapper()
-            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-    public PaymentClient() {
-        ClientHttpRequestFactorySettings settings = ClientHttpRequestFactorySettings.DEFAULTS
-                .withReadTimeout(Duration.ofSeconds(10));
-        ClientHttpRequestFactory requestFactory = ClientHttpRequestFactories.get(settings);
-        this.restClient = RestClient.builder().baseUrl(BASE_URL).requestFactory(requestFactory).build();
-    }
-
-    public void requestConfirmPayment(PaymentConfirmRequest paymentConfirmRequest) {
-        restClient.post()
-                .uri(CONFIRM_URI)
-                .contentType(MediaType.APPLICATION_JSON)
-                .header(HttpHeaders.AUTHORIZATION, createAuthorizationValue())
-                .body(paymentConfirmRequest)
-                .retrieve()
-                .onStatus(HttpStatusCode::is4xxClientError, (req, res) -> {
-                    PaymentErrorResponse errorResponse = objectMapper.readValue(res.getBody(),
-                            PaymentErrorResponse.class);
-                    throw new IllegalRequestException(errorResponse.getMessage());
-                })
-                .onStatus(HttpStatusCode::is5xxServerError, (req, res) -> {
-                    throw new InternalServerException("결제 승인 도중 알 수 없는 예외가 발생했습니다");
-                })
-                .toBodilessEntity();
-    }
-
-    private String createAuthorizationValue() {
-        byte[] encodedBytes = Base64.getEncoder().encode((secretKey + ":").getBytes(StandardCharsets.UTF_8));
-        String encodedSecretKey = new String(encodedBytes);
-
-        return AUTH_TYPE + encodedSecretKey;
-    }
+public interface PaymentClient {
+    void requestConfirmPayment(PaymentConfirmRequest paymentConfirmRequest);
 }
